requirements.txt
fastapi
uvicorn
tavily-python
google-generativeai
langchain
pydantic
python-dotenv
requests
tavily-python
pydantic-settings
langchain-groq